requirement.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.4
h5py==3.11.0
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.5.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.4.0
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.4
psutil==6.0.0
py-cpuinfo==9.0.0
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rich==13.7.1
scipy==1.14.0
seaborn==0.13.2
setuptools==72.2.0
six==1.16.0
sympy==1.13.2
tbb==2021.13.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
torch==2.3.1+cpu
torchvision==0.18.1+cpu
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.77
ultralytics-thop==2.0.0
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.44.0
wrapt==1.16.0